pyproject.toml
[tool.poetry]
name = "eventorder-backend"
version = "0.1.0"
description = ""
authors = ["Francis <francisgohqk@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
django-filter = "^24.2"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"
pynacl = "^1.5.0"
pydantic = "^2.7.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
colorlog = "^6.8.2"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
